% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveFirstWord.R
\name{RemoveFirstWord}
\alias{RemoveFirstWord}
\title{Remove First Word Function}
\usage{
RemoveFirstWord(text)
}
\arguments{
\item{text}{Takes a string as input}
}
\description{
This function allows you to remove the first word in a string
}
\examples{
RemoveFirstWord()
}
\keyword{First}
\keyword{Word}
