% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveLastWord.R
\name{RemoveLastWord}
\alias{RemoveLastWord}
\title{Remove Last Word Function}
\usage{
RemoveLastWord(text)
}
\arguments{
\item{text}{Takes a string as input}
}
\description{
This function allows you to remove the last word in a string
}
\examples{
RemoveLastWord()
}
\keyword{Last}
\keyword{Word}
